document-iq/document-iq-platform/components/ocr-adapter/pyproject.toml
[project]
name = "document-iq-platform-ocr"
version = "1.0.0"
description = "Document aggregator service for DocumentIQ platform"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "kafka-python",
    "pydantic",
    "python-dotenv",
    "redis",
    "document-iq-platform-shared"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]